Moved async stuff to world mesh creator.
