+++ Assets/GothicVR/Scenes/Bootstrap.unity
@@ -141,6 +141,7 @@ GameObject:
   - component: {fileID: 124285712}
   - component: {fileID: 124285711}
   - component: {fileID: 124285710}
+  - component: {fileID: 124285713}
   m_Layer: 0
   m_Name: Creator
   m_TagString: Untagged
@@ -269,6 +270,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 18a360ddd15d4957bfea04e16733a105, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &124285713
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 124285697}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 290769052b48463e8addfd5b2584b145, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &235819213
@@ -1013,6 +1026,8 @@ MonoBehaviour:
     m_Bits: 0
   MenuFontTag: Title
   SubtitleFontTag: IngameText
+  MeshPerFrame: 10
+  VObPerFrame: 75
 --- !u!114 &1896514221
 MonoBehaviour:
   m_ObjectHideFlags: 0
