+++ Assets/GothicVR/Scripts/Creator/Meshes/WorldMeshCreator.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 290769052b48463e8addfd5b2584b145
+timeCreated: 1690617347
